Разделил редьюсеры user и card
