+++ src/store/card/consts.js
@@ -1,0 +1,6 @@
+// Card
+export const CARD_REQUEST = 'loft-taxi/card/REQUEST';
+export const CARD_SUCCESS = 'loft-taxi/card/SUCCESS';
+export const CARD_FAILURE = 'loft-taxi/card/FAILURE';
+export const CARD_GET_INFO = 'loft-taxi/card/CARD_GET_INFO';
+export const CARD_SAVE_INFO_TO_LS = 'loft-taxi/card/CARD_SAVE_INFO_TO_LS';
