+++ src/store/user/actions.js
@@ -10,12 +10,12 @@ export const fetchUserLogin = createAction(consts.USER_LOGIN);
 export const fetchUserLogout = createAction (consts.USER_LOGOUT);
 export const fetchUserRegister = createAction(consts.USER_REGISTER);
 
-// Card
-export const fetchCardRequest = createAction(consts.CARD_REQUEST);
-export const fetchCardSuccess = createAction(consts.CARD_SUCCESS);
-export const fetchCardFailure = createAction(consts.CARD_FAILURE);
-export const fetchCardGetInfo = createAction(consts.CARD_GET_INFO);
-export const fetchCardSaveInfoToLS = createAction(consts.CARD_SAVE_INFO_TO_LS);
+// // Card
+// export const fetchCardRequest = createAction(consts.CARD_REQUEST);
+// export const fetchCardSuccess = createAction(consts.CARD_SUCCESS);
+// export const fetchCardFailure = createAction(consts.CARD_FAILURE);
+// export const fetchCardGetInfo = createAction(consts.CARD_GET_INFO);
+// export const fetchCardSaveInfoToLS = createAction(consts.CARD_SAVE_INFO_TO_LS);
 
 // Route
 export const fetchRouteRequest = createAction(consts.ROUTE_REQUEST);
