+++ src/store/user/consts.js
@@ -7,12 +7,12 @@ export const USER_REGISTER = 'loft-taxi/user/REGISTER';
 export const USER_LOGIN = 'loft-taxi/user/LOGIN';
 export const USER_LOGOUT = 'loft-taxi/user/LOGOUT';
 
-// Card
-export const CARD_REQUEST = 'loft-taxi/card/REQUEST';
-export const CARD_SUCCESS = 'loft-taxi/card/SUCCESS';
-export const CARD_FAILURE = 'loft-taxi/card/FAILURE';
-export const CARD_GET_INFO = 'loft-taxi/card/CARD_GET_INFO';
-export const CARD_SAVE_INFO_TO_LS = 'loft-taxi/card/CARD_SAVE_INFO_TO_LS';
+// // Card
+// export const CARD_REQUEST = 'loft-taxi/card/REQUEST';
+// export const CARD_SUCCESS = 'loft-taxi/card/SUCCESS';
+// export const CARD_FAILURE = 'loft-taxi/card/FAILURE';
+// export const CARD_GET_INFO = 'loft-taxi/card/CARD_GET_INFO';
+// export const CARD_SAVE_INFO_TO_LS = 'loft-taxi/card/CARD_SAVE_INFO_TO_LS';
 
 // Route
 export const ROUTE_REQUEST = 'loft-taxi/route/REQUEST';
